refactor(server): split monolithic files into focused modules
Split server.go (~1850 lines) and server_test.go (~3860 lines) into smaller,
focused files following Go convention of one test file per source file:

- auth.go/auth_test.go: authentication handlers and token logic
- handlers.go/handlers_test.go: HTTP handlers for directory listing, file operations
- file_ops.go/file_ops_test.go: file listing, sorting, template functions

Additional code quality improvements from review:
- Extract loginTemplateData struct (removes duplicate anonymous structs)
- Add getAuthUser() and getSessionMaxAge() helpers (removes duplicated logic)
- Refactor DetermineContentType() to return ContentTypeInfo struct
- Remove empty TestHandleDownloadErrors stub
