+++ .golangci.yml
@@ -49,7 +49,7 @@ linters:
     lll:
       line-length: 140
     gocyclo:
-      min-complexity: 20
+      min-complexity: 15
 
   exclusions:
     generated: lax
